worker/requirements.txt
aiohttp==3.8.4
aioredis==2.0.1
aiosignal==1.3.1
async-timeout==4.0.2
attrs==23.1.0
certifi==2023.5.7
charset-normalizer==3.1.0
frozenlist==1.3.3
idna==3.4
multidict==6.0.4
pydantic==1.10.8
python-dotenv==1.0.0
redis==3.5.3
rejson==0.5.6
requests==2.31.0
six==1.16.0
typing_extensions==4.6.3
urllib3==2.0.2
yarl==1.9.2
